Fix remaining compilation errors - database and monitoring packages
- Fixed duplicate rowsAffected declaration in slow_query_analyzer.go
- Fixed DatabaseMetrics type conversion in unified_database_monitor.go
- Fixed LegacyPerformanceMetric usage for Tags/Metadata fields
- Fixed SlowQueryThreshold access (use default if not in config)
- Fixed type mismatches (int64 vs float64)
- Fixed rows variable shadowing issue

Still have float truncation warnings in performance_report_generator.go
but those may be false positives - types look correct.

+++ internal/database/performance_monitor.go
@@ -289,7 +289,7 @@ func (pm *PerformanceMonitor) collectQueryMetrics(ctx context.Context) error {
 		totalQueries += int64(calls)
 		totalTime += totalTimeMs
 
-		if meanTimeMs > pm.config.SlowQueryThreshold.Milliseconds() {
+		if meanTimeMs > float64(pm.config.SlowQueryThreshold.Milliseconds()) {
 			slowQueries += int64(calls)
 
 			// Add to slow queries list
